% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value (optional, default = 10).}
}
\value{
The sum of x and y.
}
\description{
This function takes two parameters, x and y, and returns their sum.
If only one parameter is provided, y defaults to 10.
}
\examples{
my_add(1, 2)
my_add(1)
my_add(NA)
my_add(10, NA)
my_add("10", "20")
}
