% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_handling.R
\name{strsplit1}
\alias{strsplit1}
\title{Split string and return the first element}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character string to split.}

\item{split}{The character or string to split \code{x} on.}
}
\value{
A character vector of the split parts of \code{x}.
}
\description{
Split string and return the first element
}
\examples{
strsplit1("alfa,bravo,charlie,delta", ",")
}
