% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into substrings using a specified delimiter}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector containing the strings to split.}

\item{split}{A character vector containing the delimiter to use for splitting the strings.}
}
\value{
A character vector.
}
\description{
Split a string into substrings using a specified delimiter
}
\examples{
strsplit1("hello,world", ",")
}
